feat: move util/immutable-array-fns to root shared folder
