+++ src/app/models/form.server.model.ts
@@ -16,8 +16,8 @@ import {
   STORAGE_FORM_SETTINGS_FIELDS,
   STORAGE_PUBLIC_FORM_FIELDS,
 } from '../../../shared/constants/form'
+import { reorder } from '../../../shared/util/immutable-array-fns'
 import { MB } from '../../shared/constants'
-import { reorder } from '../../shared/util/immutable-array-fns'
 import { getApplicableIfStates } from '../../shared/util/logic'
 import {
   AuthType,
